fix(session): utilize a supervisorJob to prevent all child jobs from being cancelled if one fails
This affected probably n number of things when this was recently converted away from a ViewModel into a standalone Singleton to support SessionController (previously HomeViewModel) usage outside of just the camera screen (e.g chat). The scoping that existed in it, however, didn't associate a supervisor Job so a generic one was established. This had the adverse affects of cancelling any and all running job's within the scope if one inadvertently failed for any reason. This impacted cash link receiving as a byproduct.

Setup a supervisor job, on the IO thread, to alleviate the issue and allow child jobs to fail independently as expected.

Signed-off-by: Brandon McAnsh <git@bmcreations.dev>