fix(localfiat): cap entered amount for USDC value exchanges as well
With our switch to rounding formatted Fiat values using HALF_UP (closest), we introduced
an internal balance "capping" to the actual underlying transfer to allow bills to still render
at the entered amount (the "native amount"), while sending the min(balance, usdAmount) as the
underlying USDC value. However, this was only taken into account for custom tokens and was not applied
to straight USDC transfers as well.

Signed-off-by: Brandon McAnsh <git@bmcreations.dev>